+++ app/src/main/java/de/passwordvault/view/settings/activity_customization/SettingsCustomizationActivity.java
@@ -6,16 +6,15 @@ import androidx.lifecycle.ViewModelProvider;
 import de.passwordvault.R;
 import de.passwordvault.view.settings.dialog_darkmode.DarkmodeDialog;
 import de.passwordvault.view.settings.dialog_swipe.SwipeDialog;
-import de.passwordvault.view.settings.dialog_recently_edited.RecentlyEditedDialog;
 import de.passwordvault.view.utils.components.PasswordVaultBaseActivity;
-import de.passwordvault.view.activity_main.fragment_settings.SettingsViewModel;
+import de.passwordvault.view.settings.activity_settings.SettingsViewModel;
 
 
 /**
  * Class implements the activity which allows the user to customize the app.
  *
  * @author  Christian-2003
- * @version 3.5.2
+ * @version 3.7.0
  */
 public class SettingsCustomizationActivity extends PasswordVaultBaseActivity {
 
@@ -40,11 +39,6 @@ public class SettingsCustomizationActivity extends PasswordVaultBaseActivity {
 
         //Appearance:
         findViewById(R.id.settings_customization_appearance_darkmode_conatainer).setOnClickListener(view -> changeDarkmode());
-
-        //Home page:
-        findViewById(R.id.settings_customization_home_recentlyedited_conatainer).setOnClickListener(view -> changeNumberOfMostRecentlyEditedEntries());
-
-        //Details:
         findViewById(R.id.settings_customization_details_swipe_conatainer).setOnClickListener(view -> changeSwipeActions());
     }
 
@@ -54,15 +48,6 @@ public class SettingsCustomizationActivity extends PasswordVaultBaseActivity {
      */
     private void changeDarkmode() {
         DarkmodeDialog dialog = new DarkmodeDialog();
-        dialog.show(getSupportFragmentManager(), "");
-    }
-
-    /**
-     * Method shows the dialog to change the number of most recently edited entries on the home
-     * fragment.
-     */
-    private void changeNumberOfMostRecentlyEditedEntries() {
-        RecentlyEditedDialog dialog = new RecentlyEditedDialog();
         dialog.show(getSupportFragmentManager(), "");
     }